move CLI impl to a separate directory
