+++ src/commands/setup.rs
