+++ src/commands/shutdown.rs
